requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.1
multitasking==0.0.10
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.3
yfinance==0.1.70